worked on solving process

+++ bondrewd-builder/src/solved/mod.rs
@@ -4,10 +4,10 @@ pub mod field;
 pub mod field_set;
 
 #[must_use]
-pub fn measure_field_set_bits(fields: &[SolvedData]) -> u32 {
+pub fn measure_field_set_bits(fields: &[SolvedData]) -> usize {
     let mut l = 0;
     for f in fields {
-        l += u32::from(f.bit_length());
+        l += f.bit_length();
     }
     l
 }
